Adjust timer parameters to match real cases
So far, the essential part of the code of DAQ is finished. However,
there's two thing it can be improved in the future.

1. UART should change to DMA and use DMA interrupt to detect the
receive of UART so it won't wait until timeout every time UART is
not connected.

2. All sensors should be individual tasks and be implemented
independently in RTOS rather than run blockingly in main loop
which may cause system halt once any of the functional is broken.